Finish run footnote dependency cleanup
